Added create account functionality.
Added casing, so each username is now stored as lowercase only, and it will parse the input as lowercase before checking against database for match.

Added delete user by id endpoint, to be used with the Docs to help with testing, will delete a user from db based on ID, this is easier than opening the db with the terminal.